Image_to_Latex/bin/Debug/net6.0-windows/requirements.txt
Pillow>=10.1.0
transformers==4.34.0
Flask>=3.0.0